added a harness package
