+++ contracts/mint/tests/integration.rs
@@ -31,134 +31,10 @@ use fadroma::{
        ContractHarness, ContractEnsemble,
     },
 };
-
-pub struct Mint;
-
-impl ContractHarness for Mint {
-    // Use the method from the default implementation
-    fn init(&self, deps: &mut MockDeps, env: Env, msg: Binary) -> StdResult<InitResponse> {
-        init(
-            deps,
-            env,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-
-    fn handle(&self, deps: &mut MockDeps, env: Env, msg: Binary) -> StdResult<HandleResponse> {
-         handle(
-            deps,
-            env,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-
-    // Override with some hardcoded value for the ease of testing
-    fn query(&self, deps: &MockDeps, msg: Binary) -> StdResult<Binary> {
-         query(
-            deps,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-}
-
-pub struct MockBand;
-
-impl ContractHarness for MockBand {
-    // Use the method from the default implementation
-    fn init(&self, deps: &mut MockDeps, env: Env, msg: Binary) -> StdResult<InitResponse> {
-        mock_band::contract::init(
-            deps,
-            env,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-
-    fn handle(&self, deps: &mut MockDeps, env: Env, msg: Binary) -> StdResult<HandleResponse> {
-         mock_band::contract::handle(
-            deps,
-            env,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-
-    // Override with some hardcoded value for the ease of testing
-    fn query(&self, deps: &MockDeps, msg: Binary) -> StdResult<Binary> {
-         mock_band::contract::query(
-            deps,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-}
-
-pub struct Snip20;
-
-impl ContractHarness for Snip20 {
-    // Use the method from the default implementation
-    fn init(&self, deps: &mut MockDeps, env: Env, msg: Binary) -> StdResult<InitResponse> {
-        snip20_reference_impl::contract::init(
-            deps,
-            env,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-
-    fn handle(&self, deps: &mut MockDeps, env: Env, msg: Binary) -> StdResult<HandleResponse> {
-         snip20_reference_impl::contract::handle(
-            deps,
-            env,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-
-    // Override with some hardcoded value for the ease of testing
-    fn query(&self, deps: &MockDeps, msg: Binary) -> StdResult<Binary> {
-         snip20_reference_impl::contract::query(
-            deps,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-}
-
-pub struct Oracle;
-
-impl ContractHarness for Oracle {
-    // Use the method from the default implementation
-    fn init(&self, deps: &mut MockDeps, env: Env, msg: Binary) -> StdResult<InitResponse> {
-        oracle::contract::init(
-            deps,
-            env,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-
-    fn handle(&self, deps: &mut MockDeps, env: Env, msg: Binary) -> StdResult<HandleResponse> {
-         oracle::contract::handle(
-            deps,
-            env,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-
-    // Override with some hardcoded value for the ease of testing
-    fn query(&self, deps: &MockDeps, msg: Binary) -> StdResult<Binary> {
-         oracle::contract::query(
-            deps,
-            from_binary(&msg)?,
-            //mint::DefaultImpl,
-        )
-    }
-}
+use contract_harness::harness::mint::Mint;
+use contract_harness::harness::mock_band::MockBand;
+use contract_harness::harness::oracle::Oracle;
+use contract_harness::harness::snip20::Snip20;
 
 fn test_ensemble(offer_price: Uint128, offer_amount: Uint128, mint_price: Uint128, expected_amount: Uint128) {
