+++ packages/contract_harness/src/harness_macro.rs
@@ -1,0 +1,32 @@
+macro_rules! implement_harness {
+    ($x:ident, $s:ident) => {
+        use cosmwasm_std::{Binary, Env, from_binary, HandleResponse, InitResponse, StdResult};
+        use fadroma_ensemble::{ContractHarness, MockDeps};
+        impl ContractHarness for $x {
+            fn init(&self, deps: &mut MockDeps, env: Env, msg: Binary) -> StdResult<InitResponse> {
+                $s::contract::init(
+                    deps,
+                    env,
+                    from_binary(&msg)?,
+                )
+            }
+
+            fn handle(&self, deps: &mut MockDeps, env: Env, msg: Binary) -> StdResult<HandleResponse> {
+                $s::contract::handle(
+                    deps,
+                    env,
+                    from_binary(&msg)?
+                )
+            }
+
+            fn query(&self, deps: &MockDeps, msg: Binary) -> StdResult<Binary> {
+                $s::contract::query(
+                    deps,
+                    from_binary(&msg)?
+                )
+            }
+        }
+    }
+}
+
+pub(crate) use implement_harness;
