+++ packages/contract_harness/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod harness_macro;
+pub mod harness;
